% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{install_microbetrace}
\alias{install_microbetrace}
\title{install_microbetrace}
\usage{
install_microbetrace(dev = FALSE)
}
\arguments{
\item{dev}{}
}
\description{
install_microbetrace
}
\examples{
MicrobeTraceShiny::install_microbetrace()
}
