% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch_microbetrace}
\alias{launch_microbetrace}
\title{launch_microbetrace}
\usage{
launch_microbetrace(dev = FALSE, port)
}
\arguments{
\item{dev}{Should we launch the dev Version of MicrobeTrace?}

\item{port}{On what port should the Shiny Server Launch?}
}
\description{
launch_microbetrace
}
\examples{
MicrobeTraceShiny::launch_microbetrace()
}
